imageAnalysis/requirements.txt
sentence-transformers==3.0.1
pillow==10.4.0
Flask==3.0.3
Flask-Cors==4.0.1